refactor(recruit): abstract heading node level capping into Editor class
